requirements.txt
beautifulsoup4
discord.py
gspread
oauth2client
requests